pyproject.toml
[project]
name = "cards-for-article"
version = "0.1.0"
description = ""
authors = [
    {name = "a.dolgov",email = "dolgov_am@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "fake-useragent (>=2.2.0,<3.0.0)",
    "googlesearch-python (>=1.3.0,<2.0.0)",
    "openai (>=1.79.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
